backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.120.0
h11==0.16.0
idna==3.11
ping3==5.1.5
pydantic==2.12.3
pydantic_core==2.41.4
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
anthropic>=0.36.0
python-dotenv>=1.0.1
livekit>=1.0.7
